Split namespaces
