+++ Up2dateService/Up2dateService/Interfaces/IPackageInstaller.cs
@@ -2,7 +2,7 @@
 using System.Diagnostics;
 using Up2dateShared;
 
-namespace Up2dateService.SetupManager
+namespace Up2dateService.Interfaces
 {
     public interface IPackageInstaller
     {
